Keep bound view classes in separate modules.
